remove appcompat dependency, added context  param , corrected k-docs to reflect changes, moved file to model package
